Add lots of file caching code
