+++ Cargo.toml
@@ -18,6 +18,8 @@ futures = "0.1.25"
 tokio-fs = "0.1.4"
 log = "0.4.6"
 tokio-timer = "0.2.8"
+serde_json = "1.0"
+err-derive = "0.1.5"
 
 [dev-dependencies]
 tokio = "0.1"
